requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dateutil==2.8.2
pytz==2024.1
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
itsdangerous==2.1.2
pandas==2.2.1
pydantic==2.6.3
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-multipart==0.0.7
# Authentication dependencies - Flask-Security-Too (modern successor to Flask-Security)
flask-security-too==5.6.2
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
flask-wtf==1.2.1
wtforms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2

# API and CORS support for React frontend
flask-cors==4.0.0
flask-restful==0.3.10

# Testing dependencies
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.0.0
pytest-mock==3.12.0
httpx==0.27.0
factory-boy==3.3.0
faker==24.0.0